+++ webrtc/src/webrtc_c/api/rtc_error.cc
@@ -2,6 +2,7 @@
 
 #include <stdarg.h>
 #include <stddef.h>
+#include <cassert>
 #include <string>
 
 // WebRTC
@@ -24,11 +25,9 @@ WEBRTC_EXPORT void webrtc_RTCError_message(struct webrtc_RTCError* self,
                                            const char** out_message,
                                            size_t* out_len) {
   auto err = reinterpret_cast<webrtc::RTCError*>(self);
-  if (out_message != nullptr) {
-    *out_message = err->message();
-  }
-  if (out_len != nullptr) {
-    *out_len = std::string(err->message()).size();
-  }
+  assert(out_message != nullptr);
+  *out_message = err->message();
+  assert(out_len != nullptr);
+  *out_len = std::string(err->message()).size();
 }
 }